pyproject.toml
[project]
name = "whatagentswant"
description = "Do frontier AI models fall for the same advertising tricks as humans?"
requires-python = ">=3.12"
version = "0.1.0"
dependencies = [
    "agent-framework-core>=1.0.0rc5",
    "agent-framework-anthropic>=1.0.0b260319",
    "pandas>=2.0",
    "scipy>=1.11",
    "matplotlib>=3.8",
    "seaborn>=0.13",
    "pyyaml>=6.0",
    "python-dotenv>=1.0",
    "openpyxl>=3.1.5",
    "python-docx>=1.2.0",
]

[tool.uv]
prerelease = "if-necessary-or-explicit"
